Split ContractsTest to master, relay and rr test
